web_app/requirements.txt
pymongo
pyspellchecker
spacy
fastapi
nltk
fuzzywuzzy
Levenshtein
investigate
python-multipart
scikit-learn
nltk
uvicorn